app/docker/visitors-service/requirements.txt
Django==2.2.2
djangorestframework==3.9.4
django-cors-headers==2.4.0
mysqlclient==1.3.13
pytz==2019.1
sqlparse==0.3.0
